cleanup routing structure
