+++ packages/marketplace/ui/src/routes/__root.tsx
@@ -1,10 +1,13 @@
 import { createRootRouteWithContext, Outlet } from "@tanstack/react-router";
 import type { QueryClient } from "@tanstack/react-query";
+import { SidebarInset, SidebarProvider } from "@/components/ui/sidebar";
+import { AppSidebar } from "@/components/app-sidebar";
 
 const TanStackRouterDevtools =
   process.env.NODE_ENV === "production"
     ? () => null // Render nothing in production
     : () => null; // Render nothing in production
+
 // React.lazy(() =>
 //     import("@tanstack/router-devtools").then((res) => ({
 //       default: res.TanStackRouterDevtools,
@@ -16,8 +19,13 @@ export const Route = createRootRouteWithContext<{
 }>()({
   component: () => (
     <>
-      <Outlet />
-      <TanStackRouterDevtools />
+      <SidebarProvider>
+        <AppSidebar />
+        <SidebarInset className="bg-gray-50 min-h-screen w-full">
+          <Outlet />
+          <TanStackRouterDevtools />
+        </SidebarInset>
+      </SidebarProvider>
     </>
   ),
 });
