--- packages/marketplace/ui/src/routes/app.tsx
@@ -1,18 +1,0 @@
-import { AppSidebar } from "@/components/app-sidebar";
-import { SidebarInset, SidebarProvider } from "@/components/ui/sidebar";
-import { createFileRoute, Outlet } from "@tanstack/react-router";
-
-export const Route = createFileRoute("/app")({
-  component: RouteComponent,
-});
-
-function RouteComponent() {
-  return (
-    <SidebarProvider>
-      <AppSidebar />
-      <SidebarInset className="bg-gray-50 min-h-screen w-full">
-        <Outlet />
-      </SidebarInset>
-    </SidebarProvider>
-  );
-}
